+++ src/components/header/HeaderComponent.jsx
@@ -2,14 +2,11 @@ import { useState } from 'react'
 import MessageDropdown from '../messageDropdown/MessageDropdown'
 import './HeaderComponent.css'
 import { useHistory } from 'react-router-dom'
-import SettingsActionModal from '../settingsActionModal/SettingsActionModal'
+import ProfileMenu from './ProfileMenu'
 
 const HeaderComponent = ({ title }) => {
   const [active, setActive] = useState(false)
-  const [profileSettings, setProfileSettings] = useState(false)
   const [notification, setNotification] = useState(false)
-  const [modalActive, setModalActive] = useState(false)
-  const [settingAction, setSettingAction] = useState(null)
 
   function messageNoti () {
     setActive(!active)
@@ -26,19 +23,8 @@ const HeaderComponent = ({ title }) => {
     history.push('/messenger')
   }
 
-  function profileSettingNoti () {
-    setProfileSettings(!profileSettings)
-    setActive(false)
-  }
-
-  function clickProfileHandler (settings) {
-    setSettingAction(settings)
-    setModalActive(true)
-  }
-
   return (
     <>
-      <div>{modalActive && <SettingsActionModal setModalActive={setModalActive} settingAction={settingAction} />} </div>
       <header className='header-container'>
         <h3>{title}</h3>
         <ul>
@@ -70,17 +56,9 @@ const HeaderComponent = ({ title }) => {
               />
             )}
           </li>
-          <li onClick={() => profileSettingNoti()} className='message'>
-            <img src='/img/avatar-img.svg' alt='avatar-img' />
-            {profileSettings && (
-              <MessageDropdown
-                clickHandler={setProfileSettings}
-                clickProfileHandler={clickProfileHandler}
-                profileSettings={profileSettings}
-                message='Your settings'
-                btnName='See all notifications'
-              />
-            )}
+          
+          <li className='message'>
+           <ProfileMenu />
           </li>
         </ul>
       </header>
